.claude/worktrees/fervent-khayyam-60b5a6/src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-evaluation==1.11.0",
    "azure-ai-inference==1.0.0b9",
    "azure-ai-projects==2.0.0",
    "azure-cosmos==4.9.0",
    "azure-identity==1.24.0",
    "azure-monitor-events-extension==0.1.0",
    "azure-monitor-opentelemetry==1.8.5",
    "azure-search-documents==11.5.3",
    "fastapi==0.116.1",
    "openai==2.16.0",
    "opentelemetry-api==1.39.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.0",
    "opentelemetry-exporter-otlp-proto-http==1.39.0",
    "opentelemetry-instrumentation-fastapi==0.60b0",
    "opentelemetry-instrumentation-openai==0.46.2",
    "opentelemetry-sdk==1.39.0",
    "pytest==8.4.1",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.22",
    "uvicorn==0.35.0",
    "pylint-pydantic==0.3.5",
    "pexpect==4.9.0",
    "mcp==1.26.0",
    "werkzeug==3.1.5",
    "azure-core==1.38.0",
    "agent-framework-azure-ai==1.0.0rc4",
    "agent-framework-core==1.0.0rc4",
    "agent-framework-orchestrations==1.0.0b260311",
    "urllib3==2.6.3",
    "protobuf==5.29.6",
    "cryptography==46.0.5",
    "aiohttp==3.13.3",
    "pyasn1==0.6.2",
    "nltk==3.9.3",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.15.7",
]
